test: update tests after renaming of components
